+++ NEWS.md
@@ -36,6 +36,8 @@ Version 0.15a1 - dev
 - NEW: `Hatch.paths.outer_most_paths()` returns iterable of outer most boundary paths
 - NEW: `Hatch.paths.default_paths()` returns iterable of default boundary paths
 - NEW: `Drawing.units` property to get/set document/modelspace units
+- NEW: `ezdxf.new()` argument `units` to setup document and modelspace units and
+  $MEASUREMENT setting 
 - REMOVED: deprecated `DXFEntity.transform_to_wcs()` interface, 
   use `DXFEntity.transform(ucs.matrix)`
 - REMOVED: deprecated `Hatch.edit_boundary()` context manager,